[#17] Dynamic Device Extensions Support
- [#18] Fixed header and pointer issues first

+++ src/Rasterizer.cpp
@@ -25,8 +25,8 @@ static std::vector<glm::vec3> vertices = {
 
 static std::vector<uint32_t> indices = {0, 1, 3, 3, 2, 0, 5, 4, 6, 6, 7, 5, 1, 5, 7, 7, 3, 1, 4, 0, 2, 2, 6, 4, 3, 7, 6, 6, 2, 3, 0, 4, 5, 5, 1, 0};
 
-static const uint32_t vertexCount = static_cast<uint32_t>(vertices.size());
-static const uint32_t indexCount  = static_cast<uint32_t>(indices.size());
+static const uint32_t vertexCount = vertices.size();
+static const uint32_t indexCount  = indices.size();
 
 namespace gm {
   std::unique_ptr<Buffer> Rasterizer::vertexBuffer = nullptr;
